% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{case_and_space}
\alias{case_and_space}
\title{Remove whitespace and convert registrations to upper case}
\usage{
case_and_space(x)
}
\arguments{
\item{x}{character string}
}
\description{
Remove whitespace and convert registrations to upper case
}
\keyword{internal}
