% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{current_reg}
\alias{current_reg}
\title{Match current car registration plates (from 01 September 2001)}
\usage{
current_reg(x)
}
\arguments{
\item{x}{character vector of registrations}
}
\value{
indices of x that match current registration pattern
}
\description{
Matches all character inputs of the form 2 letters, 2 numbers, 3 letters.
}
