% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{current_reg_date}
\alias{current_reg_date}
\title{Registration date of current number plates}
\usage{
current_reg_date(x)
}
\arguments{
\item{x}{character vector of current-style registration plates}
}
\value{
the earliest possible registration date of the given reg plates
}
\description{
Registration date of current number plates
}
\keyword{internal}
