% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{dateless_reg}
\alias{dateless_reg}
\title{Match dateless car registration plates (pre 1963 - I believe)}
\usage{
dateless_reg(x)
}
\arguments{
\item{x}{character vector of registrations}
}
\value{
indices of x that match current registration pattern
}
\description{
Matches all character inputs of the forms:
  - 1 to 4 numbers, 1 to 2 letters
  - 1 to 3 numbers, 1 to 3 letters
  - 1 to 2 letters, 1 to 4 numbers
  - 1 to 3 letters, 1 to 3 numbers
}
