% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{isValidReg}
\alias{isValidReg}
\alias{isCurrentReg}
\alias{isPrefixReg}
\alias{isSuffixReg}
\alias{isDatelessReg}
\title{Match car registration plates}
\usage{
isCurrentReg(x)

isPrefixReg(x)

isSuffixReg(x)

isDatelessReg(x)

isValidReg(x)
}
\arguments{
\item{x}{character vector of registrations.}
}
\value{
A logical vector (match or not for each element of x).
}
\description{
These functions validate GB registration plates of varying formats.
}
\details{
* isCurrentReg matches current plates (from 01 September 2001) of the form 2
  letters, 2 numbers, 3 letters.
* isPrefixReg matches historic plates (01 August 1983 to 31 August 2001) with a
  1 letter prefix. These are of the form 1 letter, 1 to 3 numbers, 3 letters.
* isSuffixReg matches historic plates (February 1963 to 31 July 1983) with a
  1 letter suffix. These are of the form 3 letters, 1 to 3 numbers, 1 letter.
* isDatelessReg matches historic dateless plates (pre 1963 - I believe) of the
  following forms:
    - 1 to 4 numbers, 1 to 2 letters
    - 1 to 3 numbers, 1 to 3 letters
    - 1 to 2 letters, 1 to 4 numbers
    - 1 to 3 letters, 1 to 3 numbers
* isValidReg matches all of the above formats.
}
