% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{prefix_reg}
\alias{prefix_reg}
\title{Match car registration plates with a one letter prefix
(01 August 1983 to 31 August 2001)}
\usage{
prefix_reg(x)
}
\arguments{
\item{x}{character vector of registrations}
}
\value{
indices of x that match current registration pattern
}
\description{
Matches all character inputs of the form 1 letter, 1 to 3 numbers, 3 letters.
}
