% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{prefix_reg_date}
\alias{prefix_reg_date}
\title{Registration date of prefix style number plates}
\usage{
prefix_reg_date(x)
}
\arguments{
\item{x}{character vector of prefix-style registration plates}
}
\value{
the earliest possible registration date of the given reg plates
}
\description{
Registration date of prefix style number plates
}
\keyword{internal}
