% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{reg_match}
\alias{reg_match}
\title{Match registrations to a regex pattern}
\usage{
reg_match(x, my_regex, ...)
}
\arguments{
\item{x}{character vector of registrations}

\item{my_regex}{regex pattern you want to match}

\item{...}{additional arguments that \code{grep} can use}
}
\value{
indices of x that match the given pattern
}
\description{
Wrapper function around grep that returns the indices of the vector that
match the provided regular expression.  To help with matching whitespace and
case are ignored.
}
\keyword{internal}
