% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{regDate}
\alias{regDate}
\title{Return registration date of number plate(s)}
\usage{
regDate(x)
}
\arguments{
\item{x}{character vector of registrations}
}
\value{
The corresponding registration plate dates
}
\description{
Return registration date of number plate(s)
}
