% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{suffix_reg}
\alias{suffix_reg}
\title{Match car registration plates with a one letter suffix
(February 1963 to 31 July 1983)}
\usage{
suffix_reg(x)
}
\arguments{
\item{x}{character vector of registrations}
}
\value{
indices of x that match current registration pattern
}
\description{
Matches all character inputs of the form 3 letters, 1 to 3 numbers, 1 letter.
}
