% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{suffix_reg_date}
\alias{suffix_reg_date}
\title{Registration date of suffix style number plates}
\usage{
suffix_reg_date(x)
}
\arguments{
\item{x}{character vector of suffix-style registration plates}
}
\value{
the earliest possible registration date of the given reg plates
}
\description{
Registration date of suffix style number plates
}
\keyword{internal}
