% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid_reg}
\alias{valid_reg}
\title{Match all valid car registration plates}
\usage{
valid_reg(x)
}
\arguments{
\item{x}{character vector of registrations}
}
\value{
indices of x that match a valid registration pattern
}
\description{
Matches plates of the following format and returns their positional index
  - dateable plates (1963 onwards)
    - 2 letters, 2 numbers, 3 letters
    - 1 letter, 1 to 3 numbers, 3 letters
    - 3 letters, 1 to 3 numbers, 1 letter
}
\details{
- dateless plates (pre 1963)
    - 1 to 4 numbers, 1 to 2 letters
    - 1 to 3 numbers, 1 to 3 letters
    - 1 to 2 letters, 1 to 4 numbers
    - 1 to 3 letters, 1 to 3 numbers
}
