% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{reg_date}
\alias{reg_date}
\title{Registration date of plates(s)}
\usage{
reg_date(x)
}
\arguments{
\item{x}{character vector of registrations}
}
\value{
The corresponding registration plate dates
}
\description{
Registration date of plates(s)
}
